UBUNTU: [Packaging] Allow rust overrides
BugLink: https://bugs.launchpad.net/bugs/2084693

The rustc compiler and friends should be matching certain versions when
compiling a backported kernel. Backports of the tools do not use the
same namespace as the native versions in order to co-exist. For that
reason we need a method to pass in the exact name of the rustc, rustfmt,
and bindgen to use.

This change should be picked up by oracular and later in the main
source to allow this to be done in any derivative and backport.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: John Cabaj <john.cabaj@canonical.com>
Acked-by: Agathe Porte <agathe.porte@canonical.com>
Signed-off-by: Roxana Nicolescu <roxana.nicolescu@canonical.com>